Added script that exports tile sheets
This script reads the tilemap, VRAM and CGRAM from a save state and from that, draws a full tileset. For now, the pallette is incorrect and some tiles become corrupted, however, it should be just good enough to actually be able to work with a level. Additionally, this script only works for the first level.