Changed VRAM address
Previously, address 0x18000 was used for VRAM as it perfectly rips all the tiles (minus parallax effects), but this really only works for a few levels. The actual VRAM address is located at 0x30000, and it should work for most if not all levels now. However, reading from the actual VRAM means that some parallax will be loaded and some will not, and some tiles will have the incorrect palette depending on where in the level the save state was made.

For now, the user can create a save state and a tile sheet at multiple places in the level, and then combine them in a graphics editor to get the full correct sheet. 

For "That Old Army Game" and "Revenge of the Toys", the old address can still be used as it works for those levels and perfectly rips all front-facing tiles complete with correct colors.